+++ WonderCity-main/gui.lua
@@ -1,5 +1,5 @@
 local plugin_label = 'wonder_city'
-local plugin_version = '2.1.1'
+local plugin_version = '2.1.2'
 console.print("Lua Plugin - WonderCity - Leoric - v" .. plugin_version)
 
 local gui = {}